do not use pid() as nl_pid (portid), but let kernel assign.
Launching procd in a pid namespace on a system that has another process
with pid 1 listening for NETLINK_KOBJECT_UEVENT's fails to bind with 'Address
already in use' error.

Turns out in world of namespaces using the pid() for nl_pid is not right
anymore; we can let kernel assign a unique port id by just setting nl_pid
to 0.